skills/meeting-search/requirements.txt
chromadb
sentence-transformers
python-dotenv
